Move cli utils under cli/pkg/utils
This moves the utitility functions for CLI code to a better location.
